% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plot.R
\name{simple_plot}
\alias{simple_plot}
\title{A simple_plot Function}
\usage{
simple_plot()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to connect to the db, get the queries and plot them
}
\examples{
cat_function()


}
\keyword{cats}
